requirements.txt
Pillow==9.2.0
pillow-heif==0.10.0
streamlit==1.41.1